rustc_session: enable combine-cgu for assigner target
For now it will simplify our linking process. We use .ll files
as desired outputs and at the end we want to obtain a single .ll file.
Probably we will remove this in the future, when we will organize
linkage properly. Also we leave warnings for users about the options
that we forcibly enabled. Again when we will be sure that this is
the right way, we can ommit these warning.